+++ src/TRegx/CleanRegex/Replace/ReplacePattern.php
@@ -5,9 +5,9 @@ use TRegx\CleanRegex\Exception\NotReplacedException;
 
 interface ReplacePattern extends SpecificReplacePattern
 {
-    public function throwingOtherwise(string $exceptionClassName = NotReplacedException::class): SpecificReplacePattern;
+    public function otherwiseThrowing(string $exceptionClassName = NotReplacedException::class): SpecificReplacePattern;
 
-    public function returningOtherwise($substitute): SpecificReplacePattern;
+    public function otherwiseReturning($substitute): SpecificReplacePattern;
 
     public function otherwise(callable $substituteProducer): SpecificReplacePattern;
 }